+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Docker image
+- Footer report with number of processed files, script execution time and memory
+  consumption.
+- Docker image.
 
 ## [0.0.1] - 2023-06-12